flash-kernel installation workaround
flash-kernel failed to install in an emulated environment due to missing
support for the emulated board. Former E.L.B.E. releases ignored this
failure. In the finetuning section all nessesary configuration files are
at place and a flash-kernel can be installed without failure.
Install flash-kernel in the finetuning section.

Signed-off-by: Benedikt Spranger <b.spranger@linutronix.de>
Reviewed-by: Bastian Germann <bage@linutronix.de>