Remove lock around ELBE Container
To ensure that a running pipeline only manages the own
instance of a container a lock was used, that ensured
that only one instance of the container is running.

This prevents parallel builds using that container.
To overcome this issue every pipeline generates a
unique ID which is appended to the container name.
Therfor every pipeline can manage his own container
instance and multiple pipelines using the same
container can run in parallel.

Signed-off-by: Hartmut Eilers <eilers@linutronix.de>
Reviewed-by: Benedikt Spranger <b.spranger@linutronix.de>